+++ frontend/src/shared/components/mapIconButton/MapIconButton.tsx
