+++ frontend/src/shared/components/mapIconButton/mapIconButton.stories.ts
