+++ frontend/src/shared/components/mapIconButton/mapIconButton.styled.ts
